fixes as per the last review
