+++ mm2src/coins/lightning.rs
@@ -282,7 +282,7 @@ impl MarketCoinOps for LightningCoin {
 
     fn send_raw_tx(&self, _tx: &str) -> Box<dyn Future<Item = String, Error = String> + Send> { unimplemented!() }
     fn get_raw_tx(&self, tx: &str) -> Box<dyn Future<Item = String, Error = MmError<GetRawTransactionError>> + Send> {
-        utxo_common::get_raw_tx(&self.platform_fields.platform_coin, tx)
+        utxo_common::get_raw_tx(self.platform_fields.platform_coin.as_ref(), tx)
     }
 
     fn wait_for_confirmations(
