+++ mm2src/coins/utxo/bch.rs
@@ -1037,7 +1037,7 @@ impl MarketCoinOps for BchCoin {
     }
 
     fn get_raw_tx(&self, tx: &str) -> Box<dyn Future<Item = String, Error = MmError<GetRawTransactionError>> + Send> {
-        utxo_common::get_raw_tx(self, tx)
+        utxo_common::get_raw_tx(&self.utxo_arc, tx)
     }
 
     fn wait_for_confirmations(
